+++ src/main/java/org/irfan/library/dao/BookRepository.java
@@ -14,9 +14,9 @@ import java.util.Optional;
 
 public interface BookRepository extends JpaRepository<Book,Integer>, JpaSpecificationExecutor<Book> {
     boolean existsByTitle(String title);
-    List<Book> findAllByAuthor_Id(Long author_id);
+    List<Book> findAllByAuthor_Id(Integer author_id);
     Optional<Book> findByTitle(String title);
-    default List<Book> findBooksCustom(Optional<Long> id, Optional<String> title, Optional<Long> authorId, Optional<Long> bookTypeId){
+    default List<Book> findBooksCustom(Optional<Integer> id, Optional<String> title, Optional<Integer> authorId, Optional<Integer> bookTypeId){
         Specification<Book> spec = (root, query, cb) -> {
             List<Predicate> predicates = new ArrayList<>();
